pyproject.toml
[project]
name = "machinelearning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "huggingface-hub>=1.2.3",
    "keras-cv>=0.9.0",
    "lightning>=2.6.1",
    "matplotlib>=3.10.8",
    "monai>=1.5.2",
    "numpy>=2.4.0",
    "onnx>=1.20.1",
    "onnx2torch>=1.5.15",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pydicom>=3.0.1",
    "pylidc>=0.2.3",
    "rich>=14.2.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "tensorflow>=2.20.0",
    "tf2onnx>=1.8.4",
    "torch>=2.9.1",
    "torchmetrics>=1.8.2",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "ultralytics>=8.4.7",
]
